Move types from app.store to apps.store
